feat: view details of the data access restriction list
