+++ frontend/packages/lib-shared/api/requrls/system/process.ts
@@ -6,3 +6,4 @@ export const UpdateApprovalProcessUrl = '/approval-flow/update'; // 修改审批
 export const DeleteApprovalProcessUrl = '/approval-flow/delete'; // 删除审批流
 export const ApprovalProcessDetailUrl = '/approval-flow/get'; // 审批流详情
 export const ToggleApprovalProcessUrl = '/approval-flow/enable'; // 启用｜禁用审批流
+export const GetResourceApprovingDetailUrl = '/resource-approve/detail'; // 资源审批状态详情
