+++ frontend/packages/lib-shared/models/system/process.ts
@@ -1,7 +1,8 @@
-import { ApprovalTypeEnum, ApproverTypeEnum, ProcessStatusEnum } from '@lib/shared/enums/process';
+import { ApprovalRecordStatusEnum, ApprovalTypeEnum, ApproverTypeEnum, ProcessStatusEnum } from '@lib/shared/enums/process';
 import type { SelectedUsersItem } from './module';
 import type { FilterForm } from '@cordys/web/src/components/pure/crm-advance-filter/type';
 import type { ActionNode, ConditionBranch } from '@cordys/web/src/components/business/crm-flow/types';
+import type { UserInfo } from '@lib/shared/models/user';
 
 
 export type ProcessStatusType = Exclude<ProcessStatusEnum, ProcessStatusEnum.VOIDED>;
@@ -129,5 +130,17 @@ export interface ApprovalProcessForm {
   nodes: ApprovalProcessNode[];
   basicConfig: BasicFormParams;
   moreConfig: MoreSettingsParams;
+}
+
+
+export interface ApproverItem extends Pick<UserInfo, 'id' | 'name' | 'avatar'> {
+  approveResult: ApprovalRecordStatusEnum;
+  approveReason: string;
+}
+
+export interface ApprovalPopoverDetail {
+  resourceId: string;
+  approveStatus: ProcessStatusType;
+  approveUserList: ApproverItem[];
 }
